feat: add `SourceHandle` for proper mmap cleanup
we refcount the number of `SourceHandle`s by wrapping them in an Arc and
attaching them to `Buffer`s. This allows for the mmap to stay live for
the lifetime of the `Buffer`, calling `drop` only when the last
reference goes out of scope, which will do the `munmap`.

+++ lib/rust/hmll/examples/multi_files.rs
@@ -76,7 +76,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             print!("  [{}] Fetching {} bytes... ", i, fetch_size);
 
             let start_time = Instant::now();
-            match loader.fetch(0..fetch_size, i as i32) {
+            match loader.fetch(0..fetch_size, i) {
                 Ok(buffer) => {
                     let elapsed = start_time.elapsed();
                     total_fetch_time += elapsed.as_secs_f64();
@@ -122,7 +122,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 for (start, end) in ranges {
                     if end <= info.size {
                         print!("  [{}] Range {}..{}... ", i, start, end);
-                        match loader.fetch(start..end, i as i32) {
+                        match loader.fetch(start..end, i) {
                             Ok(buffer) => println!("✓ {} bytes", buffer.len()),
                             Err(e) => println!("✗ {}", e),
                         }
